Moved Test classes to step1, 2 package
