+++ src/test/java/study/step1/SetTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
